requirements.txt
pandas>=1.5.0
yfinance>=0.2.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0 